pyproject.toml
[project]
name = "wildfire"
version = "0.1.0"
description = "Deep learning for wildfire prediction"
requires-python = ">=3.13"
dependencies = [
    "pymupdf>=1.25.0",
    "python-dotenv>=1.2.2",
    "pyzotero>=1.6.0",
    "requests>=2.32.0",
    "zotero-cli>=0.3.0",
]
